requirements.txt
numpy>=1.21.0
Pillow>=9.0.0
opencv-python>=4.8.0
scipy
tqdm
torch>=2.2.0
torchvision>=0.17.0
diffusers
transformers
accelerate
safetensors
PyYAML
omegaconf
huggingface_hub
lama-cleaner
ultralytics
git+https://github.com/facebookresearch/sam2.git
